fix: together heder list to header
