--- src/components/HeaderList/styles.ts
@@ -1,14 +1,0 @@
-import styled from 'styled-components';
-
-export const Header = styled.header`
-  margin-top: 3rem;
-  padding-left: 8px;
-  padding-right: 8px;
-  display: flex;
-  flex-direction: row;
-
-  .title {
-    margin-left: 6rem;
-    font-size: x-large;
-  }
-`;
