--- src/layout/layoutTransaction.tsx
@@ -1,14 +1,0 @@
-import HeaderList from '@/components/HeaderList/Index';
-
-export const LayoutTransaction = ({
-  children,
-}: {
-  children: React.ReactNode;
-}) => {
-  return (
-    <>
-      <HeaderList />
-      {children}
-    </>
-  );
-};
