+++ src/routes/index.tsx
@@ -7,7 +7,6 @@ import Transactions from '../pages/Auth/Transactions';
 import Login from '../pages/Public/Login';
 import Register from '../pages/Public/Register';
 import { ProtectedRoute } from './protectedRoute';
-import { LayoutTransaction } from '@/layout/layoutTransaction';
 
 export default function Routes() {
   return (
@@ -26,9 +25,9 @@ export default function Routes() {
         path="/transactions"
         element={
           <ProtectedRoute>
-            <LayoutTransaction>
+            <Menu>
               <Transactions />
-            </LayoutTransaction>
+            </Menu>
           </ProtectedRoute>
         }
       />
